Update review in finder
